Bookstore: Securing the web application Part II
User entity has been created, and it has a functioning user repository. I also implemented user-detail service and after that i added couple of users to the database with hashed passwords (its done with bcrypt hashing algorithm). The configuration now uses the user entity classes instead of in-memory users for the authentication.